+++ RELEASE_NOTES.md
@@ -1,7 +1,10 @@
 ### Fixes
 
-- Fixes #135 - `ProductModiferConfig` was not correctly defined
+- CartsApi Missing Optional Include Parameter #145 (Thanks @jracek-chl)
+- ProductModifierValues was missing value data #150. This is a small breaking change 
+  to the [ProductModifierValue](./blob/main/src/BigCommerce/ResourceModels/Catalog/Product/ProductModifierValue.php)
+  constructor.
 
+### New Features
 
-
-
+- Implement [Site Routes API](https://developer.bigcommerce.com/api-reference/264af6ae04399-get-a-site-s-routes) #144